Network parameters refactor.
