+++ conf/node-default.properties
@@ -339,7 +339,7 @@
 ## Run with a different network
 
 # Testnet network
-node.networkParameters = brs.props.NetworkParametersTestnet
+node.network = signum.net.TestnetNetwork
 
 # Mock mining (offline and accepting any nonce as valid)
 # node.networkParameters = brs.props.NetworkParametersMock
