+++ src/brs/props/PropertyServiceImpl.java
@@ -1,6 +1,8 @@
 package brs.props;
 
 import brs.Burst;
+import signum.net.NetworkParameters;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
